Split C closure scanning
