+++ docs/PHASE_PLAN.md
@@ -106,6 +106,8 @@ checked-in docs, tests, and commits only.
   impl-block, type, behavior, and precollection work.
 - C codegen now keeps match, conditional, and controlled-loop lowering in a
   focused module, separate from statement and inline-expression emission.
+- C codegen now keeps closure definition scanning in a focused module, separate
+  from type, program, function, and global emission.
 - Resolver-backed callable type-reference validation now shares the collected
   signature/body validation helper after function, top-level method, and
   `Type.impl` method dispatch restore the resolver-owned callable key, with
